rename Bio.ts to Biography.ts
