+++ src/entity/Biography.ts
